+++ test/test_helper.rb
@@ -9,4 +9,6 @@ require 'set'
 require 'logger'
 require 'statsd-instrument'
 
+require_relative 'helpers/rubocop_helper'
+
 StatsD.logger = Logger.new(File::NULL)